requirements.txt
python-telegram-bot==20.3
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
qrcode[pil]==7.4.2
